+++ frontend-next-migration/src/features/CookieConsentV3/CookieConsentV3.module.scss
@@ -1,0 +1,187 @@
+.cookieConsentV2 {
+    position: fixed;
+    bottom: 20px;
+    left: 50%;
+    transform: translateX(-50%);
+    width: 780px;
+    padding: 24px;
+    background: var(--background);
+    border: solid var(--black);
+    border-width: var(--border-desktop);
+    border-radius: var(--border-radius-lg);
+    font-family: var(--body-font);
+    height: fit-content;
+    line-height: 1.6;
+    z-index: 1000;
+    display: flex;
+    flex-direction: column;
+    overflow: hidden;
+    box-sizing: border-box;
+    box-shadow: 8px 8px 1px #000000;
+}
+
+.cookieConsentV2 button+button {
+    margin-left: 15px;
+}
+
+.cookieConsentContent {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: 24px;
+}
+
+.cookieText {
+    flex: 1;
+}
+
+.cookieHeader {
+    margin: 0 0 20px 0;
+    color: var(--primary-color);
+    font-family: var(--font-family-title);
+    line-height: 1.2;
+    text-shadow: 0px 2px 8px rgba(0, 0, 0, 0.18);
+}
+
+.cookieConsentContentText {
+    font-size: 18px;
+    color: white;
+    line-height: 1.6;
+    font-weight: 400;
+    letter-spacing: -0.01em;
+    text-align: left;
+    font-family: 'DM Sans', sans-serif;
+    width: 432px;
+    margin-bottom: 16px;
+}
+
+.cookieImage {
+    width: 236px;
+    max-width: 236px;
+    height: auto;
+}
+
+.acceptButton {
+    background: #ffa100;
+    color: #000000;
+    padding: 12px 24px;
+    border-radius: var(--border-radius-xl);
+    font-weight: 400;
+    stroke: var(--primary-color);
+    cursor: pointer;
+    border: solid var(--black);
+    box-shadow: 2px 4px 0 var(--black) !important;
+
+    &:hover {
+        opacity: 0.9;
+        transform: scale(1.05);
+        transition: all 0.2s ease;
+    }
+}
+
+.declineButton {
+    background: var(--background);
+    color: var(--primary-color);
+    padding: 12px 24px;
+    border-radius: var(--border-radius-xl);
+    border: 1.5px solid var(--border-color);
+    stroke: var(--background);
+    box-shadow: 0 0 6px rgba(0, 0, 0, 0.6);
+    border: solid var(--black);
+    box-shadow: 2px 4px 0 var(--black) !important;
+    cursor: pointer;
+
+    &:hover {
+        opacity: 0.9;
+        transform: scale(1.05);
+        transition: all 0.2s ease;
+    }
+}
+
+.highlight {
+    color: var(--primary-color);
+    text-decoration: underline;
+}
+
+/* Responsive adjustments */
+@media (max-width: breakpoint(xs)) {
+    .cookieConsentV2 {
+        width: 90%;
+    }
+
+    .cookieConsentV2 button+button {
+        margin-left: 0px;
+        padding-top: 10px;
+    }
+
+    .cookieConsentContent {
+        flex-direction: column;
+        width: 333px;
+        padding-right: 35px;
+    }
+
+    .cookieHeader {
+        font-size: 20px;
+    }
+
+    .cookieConsentContentText {
+        font-size: 16px;
+        width: 301px;
+        height: 224px;
+        line-height: 1.5;
+        padding-left: 8px;
+    }
+
+    .cookieImage {
+        width: 164px;
+        max-width: 164px;
+        margin: 0 auto 20px auto;
+        order: -1;
+    }
+
+    .acceptButton,
+    .declineButton {
+        width: 100%;
+    }
+}
+
+@media (max-width: breakpoint(sm)) {
+    .cookieConsentV2 {
+        width: 88%;
+    }
+
+    .cookieConsentV2 button+button {
+        margin-left: 0px;
+        margin-top: 10px;
+    }
+
+    .cookieConsentContent {
+        flex-direction: column;
+        width: 333px;
+        padding-right: 35px;
+    }
+
+    .cookieHeader {
+        font-size: 20px;
+    }
+
+    .cookieConsentContentText {
+        font-size: 16px;
+        width: 301px;
+        height: 224px;
+        line-height: 1.5;
+        padding-left: 8px;
+    }
+
+    .cookieImage {
+        width: 164px;
+        max-width: 164px;
+        margin: 0 auto 20px auto;
+        order: -1;
+    }
+
+    .acceptButton,
+    .declineButton {
+        width: 100%;
+    }
+}
