+++ frontend-next-migration/src/features/CookieConsentV3/index.ts
@@ -1,0 +1,1 @@
+export { default as CookieConsentV3 } from './CookieConsentV3';
